Big Commit. Main thing: Limits
- Vector field prints commented out. Only weights are displayed.
- Added barDisplay.py. Displays bars with data taken from yarp ports. (It
  can't display decimal numbers). Also ranges must be set in the code for
  now (default 0-100)
- config.py. Added arm_limits for me. The default arm_limits is still the
  same as before. If someone also wants special limits, then take a look in
  the file.
- Debug_jointlimits.py changed to new system. Also output is compatible with
  barDisplay.
- Added follower.py. Is similar to the gravity compensation mode builtin on
the robot, but this one is integrated in our system, with this one can
position the arm in desired configurations save the  position from /vectorField/pose and use it later, check that this position
is not touching limits (barDisplay). The nice thing about this is that the
position that you get from follower are more in teractive and compatible with our system, and are "singularity aware", which
is not the case with the gravity compensation mode.
- Limits management. The system is more aware of limits and changes the
weights affecting joint usage according to this.
- Limits for the last two joints for the kuka robot. This is for now in
linInter.py, this should be move in config.py and robot.py somehow.
- Monitor distance gives also rotational distances (in degrees) to goals and
obstacles (if this makes sense)
- Nullspace aware of new limits calculation
- simstart and systemstart updated to have the barDisplay executed. (very
usefull)
- roboview still doesn't have the new limits calculation integrated, but one
angle limit got extended.